refactor: change program structure and update for new execution API
