+++ src/qibolab/instruments/qm/program/__init__.py
@@ -1,0 +1,2 @@
+from .acquisition import Acquisitions, create_acquisition
+from .instructions import program
